requirements.txt
django>=5.0,<6.0
djangorestframework
django-filter
markdown
drf-spectacular
djangorestframework-simplejwt
python-decouple
dj-database-url
gunicorn
reportlab
openpyxl
ruff
